refactor(core): own player composition types
