+++ packages/core/src/dom/index.ts
@@ -8,6 +8,7 @@ export * from './hotkey/aria';
 export * from './hotkey/coordinator';
 export * from './hotkey/hotkey';
 export * from './hotkey/hotkey-events';
+export * from './player';
 export * from './store/features';
 export * from './store/selectors';
 export * from './ui/alert-dialog';
