+++ packages/core/src/dom/media-actions.ts
@@ -1,5 +1,5 @@
-import type { AnyPlayerStore } from '@videojs/media/dom';
 import { isUndefined } from '@videojs/utils/predicate';
+import type { AnyPlayerStore } from './player';
 import { selectPlaybackRate, selectTime, selectVolume } from './store/selectors';
 
 export type MediaInputActionName = 'seekStep' | 'volumeStep' | 'speedUp' | 'speedDown';
