+++ packages/core/src/dom/store/features/tests/audio-track.test.ts
@@ -1,7 +1,7 @@
 import type { AudioTrackLike } from '@videojs/media';
-import type { PlayerTarget } from '@videojs/media/dom';
 import { createStore } from '@videojs/store';
 import { describe, expect, it } from 'vitest';
+import type { PlayerTarget } from '../../../player';
 import { audioTrackFeature } from '../audio-track';
 
 class TestAudioTrackList extends EventTarget {
