+++ packages/core/src/dom/store/features/tests/fullscreen.test.ts
@@ -1,8 +1,8 @@
-import type { PlayerTarget } from '@videojs/media/dom';
 import { HTMLVideoElementHost } from '@videojs/media/dom/video-host';
 import { createStore } from '@videojs/store';
 import type { WebKitVideoElement } from '@videojs/utils/dom';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import type { PlayerTarget } from '../../../player';
 import { createMockVideo } from '../../../tests/test-helpers';
 import { selectFullscreen } from '../../selectors';
 import { fullscreenFeature } from '../fullscreen';
