+++ packages/core/src/dom/store/features/tests/playback.test.ts
@@ -1,6 +1,6 @@
-import type { PlayerTarget } from '@videojs/media/dom';
 import { createStore } from '@videojs/store';
 import { describe, expect, it, vi } from 'vitest';
+import type { PlayerTarget } from '../../../player';
 import { createMockVideo } from '../../../tests/test-helpers';
 import { playbackFeature } from '../playback';
