+++ packages/core/src/dom/store/features/tests/quality.test.ts
@@ -1,7 +1,7 @@
 import type { VideoRenditionLike } from '@videojs/media';
-import type { PlayerTarget } from '@videojs/media/dom';
 import { createStore } from '@videojs/store';
 import { describe, expect, it } from 'vitest';
+import type { PlayerTarget } from '../../../player';
 import { qualityFeature } from '../quality';
 
 class TestRenditionList extends EventTarget {
