+++ packages/core/src/dom/store/features/tests/remote-playback.test.ts
@@ -1,6 +1,7 @@
-import type { Media, PlayerTarget } from '@videojs/media/dom';
+import type { Media } from '@videojs/media/dom';
 import type { AttachContext } from '@videojs/store';
 import { describe, expect, it, vi } from 'vitest';
+import type { PlayerTarget } from '../../../player';
 import { remotePlaybackFeature } from '../remote-playback';
 
 type RemotePlaybackState = ReturnType<typeof remotePlaybackFeature.state>;
