+++ packages/core/src/dom/store/features/tests/source.test.ts
@@ -1,6 +1,6 @@
-import type { PlayerTarget } from '@videojs/media/dom';
 import { combine, createStore } from '@videojs/store';
 import { describe, expect, it, vi } from 'vitest';
+import type { PlayerTarget } from '../../../player';
 import { createMockVideo } from '../../../tests/test-helpers';
 import { sourceFeature } from '../source';
 import { timeFeature } from '../time';
