+++ packages/core/src/dom/store/features/tests/stream-type.test.ts
@@ -1,7 +1,7 @@
 import { type MediaStreamType, MediaStreamTypes } from '@videojs/media';
-import type { PlayerTarget } from '@videojs/media/dom';
 import { createStore } from '@videojs/store';
 import { describe, expect, it } from 'vitest';
+import type { PlayerTarget } from '../../../player';
 import { createMockVideo } from '../../../tests/test-helpers';
 import { streamTypeFeature } from '../stream-type';
