+++ packages/html/src/player/context.ts
@@ -1,7 +1,7 @@
-import type { PopupGroup } from '@videojs/core/dom';
+import type { AnyPlayerStore, MediaContainer, PlayerStore, PopupGroup } from '@videojs/core/dom';
 import type { ReactiveControllerHost } from '@videojs/element';
 import { type Context, type ContextConsumer, createContext } from '@videojs/element/context';
-import type { AnyPlayerStore, Media, MediaContainer, PlayerStore } from '@videojs/media/dom';
+import type { Media } from '@videojs/media/dom';
 
 // ----------------------------------------
 // Player Context
