+++ packages/html/src/ui/controls/tests/controls-element.test.ts
@@ -1,7 +1,7 @@
 import { POPUP_HOST_ATTR } from '@videojs/core';
+import type { AnyPlayerStore } from '@videojs/core/dom';
 import { ContextProvider } from '@videojs/element/context';
 import type { MediaControlsState } from '@videojs/media';
-import type { AnyPlayerStore } from '@videojs/media/dom';
 import { createStore, flush } from '@videojs/store';
 import { afterEach, describe, expect, it, vi } from 'vitest';
