+++ packages/html/src/ui/quality-radio-group/tests/quality-radio-group-element.test.ts
@@ -1,7 +1,7 @@
+import type { AnyPlayerStore } from '@videojs/core/dom';
 import { registerI18n, resetI18nRegistry } from '@videojs/core/i18n';
 import { ContextProvider } from '@videojs/element/context';
 import type { MediaQualityState } from '@videojs/media';
-import type { AnyPlayerStore } from '@videojs/media/dom';
 import { createStore } from '@videojs/store';
 import { afterEach, describe, expect, it, vi } from 'vitest';
