+++ packages/html/src/ui/tests/media-button-element.test.ts
@@ -1,7 +1,7 @@
+import type { AnyPlayerStore, PlayerTarget } from '@videojs/core/dom';
 import { createHotkey, HOTKEY_SHORTCUT_CHANGE_EVENT, playbackFeature } from '@videojs/core/dom';
 import { registerI18n, resetI18nRegistry } from '@videojs/core/i18n';
 import { ContextProvider } from '@videojs/element/context';
-import type { AnyPlayerStore, PlayerTarget } from '@videojs/media/dom';
 import { createStore, flush } from '@videojs/store';
 import { afterEach, describe, expect, it, vi } from 'vitest';
