+++ packages/react/src/ui/hotkey/hotkey.tsx
@@ -1,7 +1,7 @@
 'use client';
 
+import type { AnyPlayerStore } from '@videojs/core/dom';
 import { createHotkey, type HotkeyActionName, isHotkeyToggleAction, resolveHotkeyAction } from '@videojs/core/dom';
-import type { AnyPlayerStore } from '@videojs/media/dom';
 import type { ReactNode } from 'react';
 import { useEffect } from 'react';
